Refactor to use RoctracerLogger
