+++ libkineto/include/GenericTraceActivity.h
@@ -64,7 +64,7 @@ class GenericTraceActivity : public ITraceActivity {
   }
 
   const ITraceActivity* linkedActivity() const override {
-    return nullptr;
+    return linked;
   }
 
   int flowType() const override {
@@ -122,6 +122,7 @@ class GenericTraceActivity : public ITraceActivity {
     uint32_t type : 4;
     uint32_t start : 1;
   } flow;
+  const ITraceActivity* linked{nullptr};
 
  private:
   const TraceSpan* traceSpan_;
